Move JSON encoding/decoding into BitwardenKit
